Actualizar Upstream y Arreglar Bugs (v1.0.3)

+++ xmcl-runtime/modrinth/service.ts
@@ -2,11 +2,13 @@ import { DownloadTask } from '@xmcl/installer'
 import { ModrinthService as IModrinthService, InstallModrinthVersionResult, InstallProjectVersionOptions, ModrinthServiceKey, ResourceDomain, getModrinthVersionFileUri, getModrinthVersionUri } from '@xmcl/runtime-api'
 import { unlink } from 'fs-extra'
 import { basename, join } from 'path'
-import { Inject, LauncherApp, LauncherAppKey } from '~/app'
+import { Inject, LauncherApp, LauncherAppKey, kTempDataPath } from '~/app'
 import { kDownloadOptions } from '~/network'
 import { AbstractService, ExposeServiceKey, Singleton } from '~/service'
 import { TaskFn, kTaskExecutor } from '~/task'
 import { ResourceService } from '../resource'
+import { isNonnull } from '~/util/object'
+import filenamify from 'filenamify'
 
 @ExposeServiceKey(ModrinthServiceKey)
 export class ModrinthService extends AbstractService implements IModrinthService {
@@ -28,7 +30,8 @@ export class ModrinthService extends AbstractService implements IModrinthService
     const isSingleFile = files.length === 1
     const resources = await Promise.all(files.map(async (file) => {
       this.log(`Try install project version file ${file.filename} ${file.url}`)
-      const destination = join(this.app.temporaryPath, basename(file.filename))
+      const getTemp = await this.app.registry.get(kTempDataPath)
+      const destination = getTemp(filenamify(basename(file.filename), { replacement: '-' }))
       const hashes = Object.entries(file.hashes)
       const urls = [file.url]
       if (version) {
@@ -76,6 +79,9 @@ export class ModrinthService extends AbstractService implements IModrinthService
           icons: icon ? [icon] : [],
         }])
 
+        if (!result) {
+          return
+        }
         result.path = result.storedPath || result.path
 
         await unlink(destination).catch(() => undefined)
@@ -98,7 +104,7 @@ export class ModrinthService extends AbstractService implements IModrinthService
 
     return {
       version,
-      resources,
+      resources: resources.filter(isNonnull),
     }
   }
 }
